+++ api/lib/domain/usecases/index.js
@@ -37,6 +37,7 @@ import * as userSavedTutorialRepository from '../../../src/devcomp/infrastructur
 import * as algorithmDataFetcherService from '../../../src/evaluation/domain/services/algorithm-methods/data-fetcher.js';
 import * as smartRandom from '../../../src/evaluation/domain/services/algorithm-methods/smart-random.js';
 import { getCompetenceLevel } from '../../../src/evaluation/domain/services/get-competence-level.js';
+import * as improvementService from '../../../src/evaluation/domain/services/improvement-service.js';
 import { pickChallengeService } from '../../../src/evaluation/domain/services/pick-challenge-service.js';
 import * as scorecardService from '../../../src/evaluation/domain/services/scorecard-service.js';
 import * as stageAndStageAcquisitionComparisonService from '../../../src/evaluation/domain/services/stages/stage-and-stage-acquisition-comparison-service.js';
@@ -107,7 +108,6 @@ import * as certificationCenterInvitationRepository from '../../../src/team/infr
 import { organizationInvitationRepository } from '../../../src/team/infrastructure/repositories/organization-invitation.repository.js';
 import { userOrgaSettingsRepository } from '../../../src/team/infrastructure/repositories/user-orga-settings-repository.js';
 import * as certificationChallengesService from '../../domain/services/certification-challenges-service.js';
-import * as improvementService from '../../domain/services/improvement-service.js';
 import * as mailService from '../../domain/services/mail-service.js';
 import * as obfuscationService from '../../domain/services/obfuscation-service.js';
 import * as passwordGenerator from '../../domain/services/password-generator.js';
@@ -255,7 +255,6 @@ const dependencies = {
   flashAssessmentResultRepository,
   frameworkRepository,
   getCompetenceLevel,
-  improvementService,
   issueReportCategoryRepository,
   juryCertificationRepository,
   juryCertificationSummaryRepository,
@@ -342,6 +341,7 @@ const dependencies = {
   writeCsvUtils,
   badgeCriteriaRepository,
   sharedSessionRepository,
+  improvementService,
 };
 
 const path = dirname(fileURLToPath(import.meta.url));
