+++ api/tests/evaluation/unit/domain/services/improvement-service_test.js
@@ -1,8 +1,8 @@
 import _ from 'lodash';
 
-import * as improvementService from '../../../../lib/domain/services/improvement-service.js';
-import { constants } from '../../../../src/shared/domain/constants.js';
-import { domainBuilder, expect, sinon } from '../../../test-helper.js';
+import * as improvementService from '../../../../../src/evaluation/domain/services/improvement-service.js';
+import { constants } from '../../../../../src/shared/domain/constants.js';
+import { domainBuilder, expect, sinon } from '../../../../test-helper.js';
 
 describe('Unit | Service | ImprovementService', function () {
   describe('#filterKnowledgeElementsIfImproving', function () {
